Feat: Separate OrderProductContent
